Eliminate alternate names for "--pass-thru".
Falling through a case is only supported for bash >= 4.0.  While that
version of bash is over 6 years old (it was released on 2009-02-20),
not everyone has caught up.

The choice was between duplicating the code, splitting it out into a
function (and duplicating function calls), and this.  Now that this
script has support for custom arguments I think --pass-thru has limited
use, and the undocumented aliases are just unnecessary bloat.